+++ datumaro/plugins/data_formats/kitti_raw/base.py
@@ -241,8 +241,8 @@ class KittiRawBase(SubsetBase):
             items[frame_id] = DatasetItem(
                 id=name,
                 subset=self._subset,
-                media=PointCloud(
-                    osp.join(self._rootdir, KittiRawPath.PCD_DIR, name + ".pcd"),
+                media=PointCloud.from_file(
+                    path=osp.join(self._rootdir, KittiRawPath.PCD_DIR, name + ".pcd"),
                     extra_images=[Image(path=image) for image in sorted(images.get(name, []))],
                 ),
                 annotations=item_desc.get("annotations"),
@@ -256,8 +256,8 @@ class KittiRawBase(SubsetBase):
             items[frame_id] = DatasetItem(
                 id=name,
                 subset=self._subset,
-                media=PointCloud(
-                    osp.join(self._rootdir, KittiRawPath.PCD_DIR, name + ".pcd"),
+                media=PointCloud.from_file(
+                    path=osp.join(self._rootdir, KittiRawPath.PCD_DIR, name + ".pcd"),
                     extra_images=[Image(path=image) for image in sorted(images.get(name, []))],
                 ),
                 attributes={"frame": int(frame_id)},